add websocket browser support
